+++ Changelog.md
@@ -43,6 +43,7 @@ Changelog [format](http://keepachangelog.com/en/1.0.0/)
   * Logging out of `Clients` is handled/retried asynchronously by `ClientManager`
 * GODT-265 Alternative Routing v2 (more resiliant to short term connection drops)
 * GODT-310 Alternative parsing of `References` header (old parsing probably malformed message IDs)
+* GODT-320 Only report the same TLS issue once every 24 hours
 
 ### Fixed
 * Use correct binary name when finding location of addcert.scpt